requirements.txt
crewai==0.80.0
crewai-tools>=0.14.0
langchain==0.3.7
langchain-groq==0.2.1
python-dotenv==1.0.0
pandas==2.2.0
openpyxl==3.1.2
requests==2.31.0
pydantic>=2.0
pyyaml==6.0.1
